+++ feature/about/src/main/java/io/github/droidkaigi/confsched2019/about/ui/bindingadapter/ViewBinding.kt
@@ -1,0 +1,9 @@
+package io.github.droidkaigi.confsched2019.about.ui.bindingadapter
+
+import android.view.View
+import androidx.databinding.BindingAdapter
+
+@BindingAdapter(value = ["visibleGone"])
+fun showHide(view: View, show: Boolean) {
+    view.visibility = if (show) View.VISIBLE else View.GONE
+}
