+++ feature/about/src/main/res/values/strings.xml
@@ -3,10 +3,18 @@
     <string name="about_what_is_droidkaigi">What is DroidKaigi?</string>
     <string name="about_description_droidkaigi">DroidKaigi is a conference tailored for Android developers. It\'s scheduled to take place on the 7th and 8th of February 2019.</string>
     <string name="about_access_to_place">Directions</string>
+    <string name="about_wifi">WiFi</string>
+    <string name="about_wifi_summary" translatable="false">DroidKaigi2019 / thankyou-iridge</string>
+    <string name="about_add_wifi">Add to Device</string>
     <string name="about_check_map">Open with Maps</string>
     <string name="about_staff_list">Committee Members</string>
     <string name="about_check">See Details</string>
     <string name="about_privacy_policy">Privacy Policy</string>
     <string name="about_license">Open Source Licenses</string>
     <string name="about_app_version">Version</string>
+
+    <string name="wifi_ssid" translatable="false">DroidKaigi2019</string>
+    <string name="wifi_password" translatable="false">thankyou-iridge</string>
+    <string name="wifi_failed_to_register_message_so_copied">Copied the password to Clipboard instead. The SSID is %1$s</string>
+    <string name="wifi_failed_to_register_message">Failed to register. SSID is %1$s, password is %2$s</string>
 </resources>
