+++ feature/settings/build.gradle
@@ -23,7 +23,7 @@ dependencies {
     implementation Dep.AndroidX.appCompat
     implementation Dep.AndroidX.constraint
     implementation Dep.AndroidX.recyclerView
-    api Dep.AndroidX.preference
+    implementation Dep.AndroidX.preference
     api Dep.AndroidX.emoji
     implementation Dep.AndroidX.design
