--- feature/settings/src/main/res/drawable/ic_wifi.xml
@@ -1,4 +1,0 @@
-<vector android:height="18dp" android:viewportHeight="48.26"
-    android:viewportWidth="68.27" android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
-    <path android:fillColor="#ed4239" android:pathData="M0,14.13l6.21,6.2a39.5,39.5 0,0 1,55.85 0l6.21,-6.2A48.3,48.3 0,0 0,0 14.13ZM24.82,39l9.31,9.31L43.44,39A13.15,13.15 0,0 0,24.82 39ZM12.41,26.54l6.21,6.2a22,22 0,0 1,31 0l6.2,-6.2A30.75,30.75 0,0 0,12.41 26.54Z"/>
-</vector>
