+++ feature/settings/src/main/res/values/strings.xml
@@ -1,11 +1,3 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <string name="pref_key_add_wifi" translatable="false">key_add_wifi</string>
-
-    <string name="title_add_wifi">Add Wifi setting</string>
-    <string name="summary_add_wifi" translatable="false">Sponsored by iRidge</string>
-    <string name="wifi_ssid" translatable="false">DroidKaigi2019</string>
-    <string name="wifi_password" translatable="false">thankyou-iridge</string>
-    <string name="wifi_failed_to_register_message_so_copied">Copied the password to Clipboard instead. The SSID is %1$s</string>
-    <string name="wifi_failed_to_register_message">Failed to register. SSID is %1$s, password is %2$s</string>
 </resources>
