+++ feature/settings/src/main/res/xml/preferences.xml
@@ -1,10 +1,4 @@
 <?xml version="1.0" encoding="utf-8"?>
 <PreferenceScreen
     xmlns:android="http://schemas.android.com/apk/res/android">
-    <Preference
-        android:icon="@drawable/ic_wifi"
-        android:key="@string/pref_key_add_wifi"
-        android:summary="@string/summary_add_wifi"
-        android:title="@string/title_add_wifi"
-        />
 </PreferenceScreen>
